fix: replacing the pending state correctly with keys
